pyproject.toml
[project]
name = "git-commit-extractor"
version = "0.1.0"
description = ""
authors = [
    {name = "Jakub Hána"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gitpython (>=3.1.45,<4.0.0)",
    "delegator-py (>=0.1.1,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
